consolodated tests and functions into single files
