--- test/get.test.js
@@ -1,12 +1,0 @@
-const assert = require('assert');
-const getTest = require('../lib/get.js');
-
-describe('get reads the contents of a record', function() {
-    it('read contents of record', function(done) {
-        getTest.get('pets/dogs', 'ABC123', function(err, data) {
-            assert.deepEqual(data, { database: 'pets', collection: 'dogs', owner: 'Mike', petName: 'Tux', _id: 'ABC123' });
-            done();
-        });
-
-    });
-});
